old/modelgenerator/requirements.txt
requests>=2.31.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
google-cloud-pubsub==2.28.0
google-cloud-storage==2.19.0
pathlib==1.0.1